Combined multiple sources uglyly (has issues)
There's an issue with Teams being removed despite being (triple) commented out in the remote file and not appearing in any of the lists. I think this copy of Teams is actually the 'personal' teams and is probably best if it's bulk removed and then the Teams Bootstrapper installed if required; but I really don't know.